+++ pkg/query/model.go
@@ -29,10 +29,11 @@ type Query struct {
 	URL                 string                 `json:"url"`
 	URLOptions          URLOptions             `json:"url_options"`
 	Data                string                 `json:"data"`
-	Parser              string                 `json:"parser"` // 'simple' | 'backend' | 'uql' | 'groq'
+	Parser              string                 `json:"parser"` // 'simple' | 'backend' | 'sqlite' | 'uql' | 'groq'
 	SummarizeExpression string                 `json:"summarizeExpression"`
 	UQL                 string                 `json:"uql"`
 	GROQ                string                 `json:"groq"`
+	SQLiteQuery         string                 `json:"sqlite_query"`
 	CSVOptions          InfinityCSVOptions     `json:"csv_options"`
 	JSONOptions         InfinityJSONOptions    `json:"json_options"`
 	RootSelector        string                 `json:"root_selector"`
@@ -145,12 +146,12 @@ func ApplyDefaultsToQuery(query Query) Query {
 	return query
 }
 
-func LoadQuery(backendQuery backend.DataQuery) (Query, error) {
+func LoadQuery(backendQuery backend.DataQuery, pluginContext backend.PluginContext) (Query, error) {
 	var query Query
 	err := json.Unmarshal(backendQuery.JSON, &query)
 	if err != nil {
 		return query, fmt.Errorf("error while parsing the query json. %s", err.Error())
 	}
 	query = ApplyDefaultsToQuery(query)
-	return ApplyMacros(query, backendQuery.TimeRange)
+	return ApplyMacros(query, backendQuery.TimeRange, pluginContext)
 }