+++ pkg/query/model_test.go
@@ -55,6 +55,7 @@ func TestLoadQuery(t *testing.T) {
 				},
 				"uql" 		: "my-uql-query",
 				"groq"		: "my-groq-query",
+				"sqlite_query"  : "select foo from bar",
 				"expression" 	: "my-expression",
 				"seriesCount" 	: 20,
 				"alias" 		: "my-alias",
@@ -96,6 +97,7 @@ func TestLoadQuery(t *testing.T) {
 				},
 				UQL:           "my-uql-query",
 				GROQ:          "my-groq-query",
+				SQLiteQuery:   "select foo from bar",
 				Expression:    "my-expression",
 				SeriesCount:   20,
 				Alias:         "my-alias",
@@ -106,7 +108,7 @@ func TestLoadQuery(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			q := &backend.DataQuery{JSON: []byte(tt.queryJSON)}
-			got, err := querySrv.LoadQuery(*q)
+			got, err := querySrv.LoadQuery(*q, backend.PluginContext{})
 			if tt.wantErr != nil {
 				require.NotNil(t, err)
 				assert.Equal(t, tt.wantErr, err)